Estavel/old/Chunk_1_OK/requirements.txt
langchain-community
pymongo
pytesseract
pdf2image
unstructured
python-docx
pdfminer.six
python-dotenv
psycopg2-binary
requests